requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
google-generativeai==0.8.5
groq==0.32.0
requests==2.31.0
rapidfuzz==3.9.7
numpy==1.26.4
pandas==2.2.3
openpyxl==3.1.5
Werkzeug==3.1.3
python-dotenv==1.0.1
PyMySQL==1.1.1
